Take advantage of the custom template functions in the main config rendering
